feat: add type guard for refresh endpoint rejected action
This commit introduces a new type guard,
`refreshEndpointRejectedAction`, to identify rejected actions from the
refresh token endpoint. This is necessary to prevent the automated
refresh flow from being triggered incorrectly when the refresh token
itself has expired.

The type guard is implemented using `isAnyOf` matcher utility from
Redux Toolkit and checks if the action matches the `matchRejected`
case of the `refreshTokens` endpoint in the `authApiSlice`.