feat: notify user on token refresh failure
Adds a toast notification within `customizedBaseQueryWithReauth` to
inform the user when token refresh fails.

This notification appears just before the user is logged out and
redirected to the login page, providing context for the session
termination.

The toast has been placed here along with redirection because it is
a good practice to notify user and navigate them to login page on
session expiry.
We have a listener middleware in place for handling this but this
particular `refresh` endpoint is not a part of it as the
`customizedBaseQueryWithReauth` though handles reauth, does not return
any signal for the failure of the same and just logs errors and releases
the mutex, hence the result being returned being that of a pending
api call.