feat: scroll to new message in editor
This commit updates the `MarkdownTextEditor` to automatically scroll the
user to the newly created message.

The component now accepts a `messageViewType` prop, which indicates the
current sort order of the message list. Based on this, it scrolls to either
the top (for 'newest' sort) or the bottom of the page to make the new message visible.

This improves the user experience by providing immediate feedback after a
message is sent. The component also now accepts a `className` for more flexible styling.